requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
supabase>=2.0.0
Pillow
python-docx
pypdf
